+++ changelog.md
@@ -20,6 +20,7 @@ Codion Change Log
 - FilteredComboBoxModel.sortVisibleItems() bug fixed, now takes the null item into account.
 - Controls.Builder.defaults(stopAt) now includes the specified stopAt control instead of stopping before it.
 - DialogBuilder, FrameBuilder.titleProvider() renamed title().
+- LookAndFeelProvider.addLookAndFeelProvider() renamed addLookAndFeel().
 ### is.codion.swing.common.ui
 - ProgressDialog.Builder, ProgressWorkerDialogBuilder.controls() replaced with control().
 - ProgressDialog bug fixed, no longer adds an empty south panel when no controls are specified.